+++ app/src/main/java/com/ama/algorithmmanagement/activity/kDefault/MyChildCommentActivity.kt
@@ -2,7 +2,7 @@ package com.ama.algorithmmanagement.activity.kDefault
 
 import android.os.Bundle
 import androidx.lifecycle.ViewModelProvider
-import com.ama.algorithmmanagement.adapter.KChildCommentsAdapter
+import com.ama.algorithmmanagement.presentation.vpdetail.adapter.KChildCommentsAdapter
 import com.ama.algorithmmanagement.application.AMAApplication
 import com.ama.algorithmmanagement.domain.base.BaseViewModelFactory
 import com.ama.algorithmmanagement.domain.base.KBaseActivity
