+++ app/src/main/java/com/ama/algorithmmanagement/presentation/signup/fragment/ConnectBOJAccountFragment.kt
@@ -1,4 +1,4 @@
-package com.ama.algorithmmanagement.Fragment
+package com.ama.algorithmmanagement.presentation.signup.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -12,15 +12,15 @@ import com.ama.algorithmmanagement.domain.base.BaseViewModelFactory
 import com.ama.algorithmmanagement.domain.base.KBaseFragment
 import com.ama.algorithmmanagement.R
 import com.ama.algorithmmanagement.data.repositories.RepositoryLocator
-import com.ama.algorithmmanagement.databinding.FragmentRegisterFinalBinding
-import com.ama.algorithmmanagement.viewmodel.kDefault.KSignUpViewModel
+import com.ama.algorithmmanagement.databinding.FragmentConnectBojAccountBinding
+import com.ama.algorithmmanagement.presentation.signup.view_model.KSignUpViewModel
 
 /**
  * author : manyong Han
- * summary : 회원가입 마지막 화면 프래그먼트 (아이디, 비밀번호 입력)
+ * summary : 백준 연동하는 화면 프래그먼트
  */
 
-class RegisterFinalFragment : KBaseFragment<FragmentRegisterFinalBinding>(R.layout.fragment_register_final) {
+class ConnectBOJAccountFragment() : KBaseFragment<FragmentConnectBojAccountBinding>(R.layout.fragment_connect_boj_account) {
 
     private lateinit var signUpViewModel: KSignUpViewModel
 
@@ -37,8 +37,18 @@ class RegisterFinalFragment : KBaseFragment<FragmentRegisterFinalBinding>(R.layo
         return super.onCreateView(inflater, container, savedInstanceState)
     }
 
+    // 뷰모델을 부모꺼 사용하면서 값 변경해주는걸 알려주기만 하면 아래처럼 안해도됨
+    // 때문에 뷰모델을 부모액티비티꺼를 사용한다? => 옵저빙이 액티비티에서 실행되면 된다.
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         binding.viewModel = signUpViewModel
+
+        signUpViewModel.isConnectSuccess.observe(viewLifecycleOwner) {
+            if(it) {
+                binding.connectNextBtn.visibility = View.VISIBLE
+                signUpViewModel.isConnectSuccess.value = false
+                signUpViewModel.isMoveToWebView.value = false
+            }
+        }
     }
 
     // requireActivity() 는 getActivity() 가 null 일 경우에 IllegalAccessException 을 던지기 때문에
