+++ app/src/main/java/com/ama/algorithmmanagement/presentation/vpdetail/adapter/KChildCommentsAdapter.kt
@@ -1,9 +1,9 @@
-package com.ama.algorithmmanagement.adapter
+package com.ama.algorithmmanagement.presentation.vpdetail.adapter
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.ama.algorithmmanagement.viewHolder.ChildCommentViewHolder
 import com.ama.algorithmmanagement.data.model.ChildCommentInfo
+import com.ama.algorithmmanagement.presentation.vpdetail.view_holder.ChildCommentViewHolder
 
 /**
  * author : manyong Han
