+++ app/src/main/java/com/ama/algorithmmanagement/presentation/vpdetail/adapter/KCommentsAdapter.kt
@@ -1,9 +1,10 @@
-package com.ama.algorithmmanagement.adapter
+package com.ama.algorithmmanagement.presentation.vpdetail.adapter
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.ama.algorithmmanagement.viewHolder.CommentViewHolder
 import com.ama.algorithmmanagement.data.model.CommentInfo
+import com.ama.algorithmmanagement.presentation.vpdetail.view_holder.CommentViewHolder
+
 /**
  * author : manyong Han
  * summary : 댓글 어댑터
