+++ app/src/main/java/com/ama/algorithmmanagement/presentation/vpdetail/view_holder/CommentViewHolder.kt
@@ -1,4 +1,4 @@
-package com.ama.algorithmmanagement.viewHolder
+package com.ama.algorithmmanagement.presentation.vpdetail.view_holder
 
 import android.view.ViewGroup
 import com.ama.algorithmmanagement.domain.base.KBaseViewHolder
