+++ app/src/main/java/com/ama/algorithmmanagement/presentation/vpdetail/view_model/KViewProblemDetailViewModel.kt
@@ -1,4 +1,4 @@
-package com.ama.algorithmmanagement.viewmodel.kDefault
+package com.ama.algorithmmanagement.presentation.vpdetail.view_model
 
 import androidx.databinding.ObservableArrayList
 import androidx.lifecycle.*
