+++ app/src/main/java/com/ama/algorithmmanagement/utils/BindingAdapterUtils.kt
@@ -12,6 +12,8 @@ import com.ama.algorithmmanagement.Fragment.SolvedProblemViewPagerFragment
 import com.ama.algorithmmanagement.data.model.*
 import com.ama.algorithmmanagement.presentation.main.adapter.KRetryProblemsAdapter
 import com.ama.algorithmmanagement.presentation.main.adapter.KUserDateInfoAdapter
+import com.ama.algorithmmanagement.presentation.vpdetail.adapter.KChildCommentsAdapter
+import com.ama.algorithmmanagement.presentation.vpdetail.adapter.KCommentsAdapter
 import com.github.mikephil.charting.charts.HorizontalBarChart
 import com.github.mikephil.charting.charts.PieChart
 import com.github.mikephil.charting.components.XAxis
