+++ app/src/main/res/layout/activity_login.xml
@@ -6,13 +6,13 @@
     <data>
         <variable
             name="viewModel"
-            type="com.ama.algorithmmanagement.viewmodel.kDefault.KRLoginViewModel" />
+            type="com.ama.algorithmmanagement.presentation.login.view_model.KRLoginViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.kDefault.KRLoginActivity">
+        tools:context=".presentation.login.activity.KRLoginActivity">
 
         <TextView
             android:id="@+id/login_act_logo"
